+++ waspc/ChangeLog.md
@@ -29,6 +29,7 @@
 
 ### 🐞 Bug fixes
 
+- When an env file defines the same variable more than once, Wasp now uses the last definition (instead of previous first). ([#4895](https://github.com/wasp-lang/wasp/pull/4895))
 - Wasp's compiler now properly escapes user input in the code it generates. ([#4855](https://github.com/wasp-lang/wasp/pull/4855))
 
 - Fixed a bug that created Railway databases without a persistent volume when using `--db-image`. Existing databases are unchanged. ([#4647](https://github.com/wasp-lang/wasp/pull/4647))